+++ api/dictionnary.cjs
@@ -1,7 +1,5 @@
 'use strict'
-
 const fs = require('fs')
-const letterReplacements = require('./letter_replacements.cjs')
 
 const filename = 'api/dictionnary_2021_06_07.json'
 const dictionnary = JSON.parse(fs.readFileSync(filename))
@@ -18,28 +16,11 @@ const letters = dictionnary.reduce((letters, definition) => {
 
 console.log(`${Object.keys(letters).length} letters loaded`)
 
-const normalizePersian = (word) => {
-  var newWord = word
-
-  for (var letter in letterReplacements) {
-    const replacements = letterReplacements[letter].join('|')
-    const regex = new RegExp(`(${replacements})`, 'g')
-    newWord = newWord.normalize().replace(regex, letter)
-  }
-
-  return newWord
-}
+const farsi_search = require('./farsi_search.cjs')(dictionnary)
 
 module.exports = {
   async find (word) {
-    const foreignResults = dictionnary.filter((definition) => {
-      return normalizePersian(definition.foreign_word).includes(normalizePersian(word))
-    })
-    const persianResults = dictionnary.filter((definition) => {
-      return normalizePersian(definition.persian_word).includes(normalizePersian(word))
-    })
-
-    var definitions = [...foreignResults, ...persianResults]
+    var definitions = farsi_search.search_without_accents(word)
     const nbDefinitions = definitions.length
 
     return {
